+++ src/controllers/sse/sse.service.ts
@@ -4,7 +4,7 @@ import { Inject, Injectable } from '@nestjs/common';
 import { Request, Response } from 'express';
 
 import { McpServerOptions } from '../../interfaces/mcp-server-options.interface';
-import { McpLoggerService } from '../../registry/mcp-logger.service';
+import { McpLoggerService } from '../../registry/logger.service';
 import { RegistryService } from '../../registry/registry.service';
 
 @Injectable()
