+++ src/registry/registry.service.ts
@@ -15,8 +15,9 @@ import {
   ToolOptions,
 } from '../interfaces/capabilities.interface';
 
+import { MCP_RESOLVER } from '../decorators';
 import { DiscoveryService } from './discovery.service';
-import { McpLoggerService } from './mcp-logger.service';
+import { McpLoggerService } from './logger.service';
 
 @Injectable()
 export class RegistryService {
@@ -36,67 +37,52 @@ export class RegistryService {
     this.logger.log('MCP capabilities registration completed.', 'registry');
   }
 
-  private registerResources(server: McpServer): void {
-    this.logger.log('Starting resources registration...', 'resources');
+  private wrappedHandler<TArgs extends unknown[], TResult>(
+    handler: (...args: TArgs) => TResult,
+    instance: object,
+  ): (...args: TArgs) => TResult {
+    const isResolver = Reflect.hasMetadata(MCP_RESOLVER, instance.constructor);
 
+    if (!isResolver) {
+      throw new Error(
+        `Class "${instance.constructor.name}" must be decorated with @Resolver to use @Prompt, @Tool, or @Resource.`,
+      );
+    }
+
+    return (...args: TArgs) => handler(...args);
+  }
+
+  private registerResources(server: McpServer): void {
     const resourceMethods =
       this.discoveryService.getAllMethodsWithMetadata<ResourceOptions>(
         MCP_RESOURCE,
       );
-
-    this.logger.log(
-      `Found ${resourceMethods.length} methods with @Resource decorator`,
-      'resources',
-    );
-
     for (const method of resourceMethods) {
       const { metadata, handler } = method;
 
-      this.logger.verbose(
-        `Processing resource: ${metadata?.name || 'unnamed'}`,
+      this.logger.log(
+        `Resource "${metadata?.name || 'unnamed'}" found.`,
         'resources',
       );
 
       try {
-        // Handle all possible resource registration formats:
-        // 1. (name, template, metadata, cb) - template with metadata
-        // 2. (name, template, cb) - template without metadata
-        // 3. (name, uri, metadata, cb) - URI with metadata
-        // 4. (name, uri, cb) - URI without metadata
-
         if ('template' in metadata) {
-          this.logger.log(
-            `Registering resource "${metadata.name}" with template: ${metadata.template}`,
-            'resources',
-          );
-
           if ('metadata' in metadata) {
-            // Case 1: Template with metadata
             server.resource(
               metadata.name,
               new ResourceTemplate(metadata.template, { list: undefined }),
               metadata.metadata,
               handler,
             );
+          } else {
+            server.resource(
+              metadata.name,
+              new ResourceTemplate(metadata.template, { list: undefined }),
+              handler,
+            );
           }
-
-          this.logger.log('Resource without metadata', 'resources');
-
-          // Case 2: Template without metadata
-          console.log(metadata.template);
-          server.resource(
-            metadata.name,
-            new ResourceTemplate(metadata.template, { list: undefined }),
-            handler,
-          );
         } else if ('uri' in metadata) {
-          this.logger.log(
-            `Registering resource "${metadata.name}" with URI: ${metadata.uri}`,
-            'resources',
-          );
-
           if ('metadata' in metadata) {
-            // Case 3: URI with metadata
             server.resource(
               metadata.name,
               metadata.uri,
@@ -104,15 +90,9 @@ export class RegistryService {
               handler,
             );
           } else {
-            // Case 4: URI without metadata
             server.resource(metadata.name, metadata.uri, handler);
           }
         }
-
-        this.logger.log(
-          `Resource "${metadata.name}" successfully registered`,
-          'resources',
-        );
       } catch (error) {
         this.logger.error(
           `Error registering resource ${metadata.name}: ${error}`,
@@ -131,31 +111,20 @@ export class RegistryService {
   }
 
   private registerPrompts(server: McpServer): void {
-    this.logger.log('Starting prompts registration...', 'prompts');
-
     const promptMethods =
       this.discoveryService.getAllMethodsWithMetadata<PromptOptions>(
         MCP_PROMPT,
       );
-
-    this.logger.log(
-      `Found ${promptMethods.length} methods with @Prompt decorator`,
-      'prompts',
-    );
-
     for (const method of promptMethods) {
       const { metadata, handler } = method;
 
-      this.logger.verbose(
-        `Processing prompt: ${metadata?.name || 'unnamed'}`,
+      this.logger.log(
+        `Prompt "${metadata?.name || 'unnamed'}" found.`,
         'prompts',
       );
 
       try {
-        this.logger.log(`Registering prompt "${metadata.name}"`, 'prompts');
-
         if ('description' in metadata && 'argsSchema' in metadata) {
-          // Case 1: (name, uri, metadata, cb)
           server.prompt(
             metadata.name,
             metadata.description,
@@ -163,20 +132,12 @@ export class RegistryService {
             handler,
           );
         } else if ('argsSchema' in metadata) {
-          // Case 2: (name, argsSchema, cb)
           server.prompt(metadata.name, metadata.argsSchema, handler);
         } else if ('description' in metadata) {
-          // Case 3: (name, description, cb)
           server.prompt(metadata.name, metadata.description, handler);
         } else {
-          // Case 4: (name, cb)
           server.prompt(metadata.name, handler);
         }
-
-        this.logger.log(
-          `Prompt "${metadata.name}" successfully registered`,
-          'prompts',
-        );
       } catch (error) {
         this.logger.error(
           `Error registering prompt ${metadata.name}: ${error}`,
@@ -195,76 +156,50 @@ export class RegistryService {
   }
 
   private registerTools(server: McpServer): void {
-    this.logger.log('Starting tools registration...', 'tools');
-
     const toolMethods =
       this.discoveryService.getAllMethodsWithMetadata<ToolOptions>(MCP_TOOL);
 
-    this.logger.log(
-      `Found ${toolMethods.length} methods with @Tool decorator`,
-      'tools',
-    );
-
     for (const method of toolMethods) {
-      const { metadata, handler } = method;
+      const { metadata, handler, instance } = method;
 
-      this.logger.verbose(
-        `Processing tool: ${metadata?.name || 'unnamed'}`,
-        'tools',
-      );
+      this.logger.log(`Tool "${metadata?.name || 'unnamed'}" found.`, 'tools');
 
-      if (metadata && metadata.name && handler) {
-        try {
-          this.logger.log(`Registering tool "${metadata.name}"`, 'tools');
-
-          // Handle all possible tool registration formats:
-          // 1. (name, description, paramSchema, cb)
-          // 2. (name, paramSchema, cb)
-          // 3. (name, description, cb)
-          // 4. (name, cb)
-
-          if ('description' in metadata && 'paramSchema' in metadata) {
-            // Case 1: (name, description, paramSchema, cb)
-            server.tool(
-              metadata.name,
-              metadata.description,
-              metadata.paramSchema,
-              handler,
-            );
-          } else if ('paramSchema' in metadata) {
-            // Case 2: (name, paramSchema, cb)
-            server.tool(metadata.name, metadata.paramSchema, handler);
-          } else if ('description' in metadata) {
-            // Case 3: (name, description, cb)
-            server.tool(metadata.name, metadata.description, handler);
-          } else {
-            // Case 4: (name, cb)
-            server.tool(metadata.name, handler);
-          }
-          this.logger.log(
-            `Tool "${metadata.name}" successfully registered`,
-            'tools',
+      try {
+        if ('description' in metadata && 'paramSchema' in metadata) {
+          server.tool(
+            metadata.name,
+            metadata.description,
+            metadata.paramSchema,
+            this.wrappedHandler(handler, instance),
           );
-        } catch (error) {
+        } else if ('paramSchema' in metadata) {
+          server.tool(
+            metadata.name,
+            metadata.paramSchema,
+            this.wrappedHandler(handler, instance),
+          );
+        } else if ('description' in metadata) {
+          server.tool(
+            metadata.name,
+            metadata.description,
+            this.wrappedHandler(handler, instance),
+          );
+        } else {
+          server.tool(metadata.name, this.wrappedHandler(handler, instance));
+        }
+      } catch (error) {
+        this.logger.error(
+          `Error registering tool ${metadata.name}: ${error}`,
+          undefined,
+          'tools',
+        );
+        if (error && typeof error === 'object' && 'stack' in error) {
           this.logger.error(
-            `Error registering tool ${metadata.name}: ${error}`,
+            `Stack trace: ${(error as Error).stack}`,
             undefined,
             'tools',
           );
-
-          if (error && typeof error === 'object' && 'stack' in error) {
-            this.logger.error(
-              `Stack trace: ${(error as Error).stack}`,
-              undefined,
-              'tools',
-            );
-          }
         }
-      } else {
-        this.logger.warn(
-          `Skipping tool with incomplete metadata: ${JSON.stringify({ name: metadata?.name, hasHandler: !!handler })}`,
-          'tools',
-        );
       }
     }
   }
